requirements.txt
pandas
mysql-connector-python
sqlalchemy
python-dotenv
plotly
matplotlib
kaleido
requests    
numpy
scipy
